replace StatCursor with TypedCursor
